fixing project issues

+++ Zenith.Colour/ColourRGB.cs
@@ -2,16 +2,16 @@
 
 namespace Zenith.Colour
 {
-	public record struct ColourRGB(int R, int G, int B) : IVector3<int>
+	public record struct ColourRGB(float R, float G, float B) : IVector3<float>
 	{
 		// IVector3
-		public int X { get => R; set => R = value; }
-		public int Y { get => G; set => G = value; }
-		public int Z { get => B; set => B = value; }
+		public float X { get => R; set => R = value; }
+		public float Y { get => G; set => G = value; }
+		public float Z { get => B; set => B = value; }
 
 		public ColourHSB AsHSB()
 			=> ColourSpaceConverter.RGBtoHSB(this);
 
-		public static readonly ColourRGB None = new() { R = -1, G = -1, B = -1 };
+		public static readonly ColourRGB None = new() { R = -1f, G = -1f, B = -1f };
 	}
 }
