+++ Zenith.Drawing/ColourExtensions.cs
@@ -6,9 +6,9 @@ namespace Zenith.System.Drawing
 	public static class ColourExtensions
 	{
 		public static ColourRGB ToColourRGB(this Color c)
-			=> new() { R = (int)(c.R / 255f), G = (int)(c.G / 255f), B = (int)(c.B / 255f) };
+			=> new() { R = c.R / 255f, G = c.G / 255f, B = c.B / 255f };
 
 		public static Color ToSystemColor(this ColourRGB ColourRGB)
-			=> Color.FromArgb(ColourRGB.R * 255, ColourRGB.G * 255, ColourRGB.B * 255);
+			=> Color.FromArgb((byte)(ColourRGB.R * 255), (byte)(ColourRGB.G * 255), (byte)(ColourRGB.B * 255));
 	}
 }
